+++ lightly_studio/e2e-tests/index_captions.py
@@ -12,7 +12,7 @@ import lightly_studio as ls
 ls.db_manager.connect(cleanup_existing=True)
 
 # Create a DatasetLoader instance
-dataset = ls.Dataset.create()
+dataset = ls.ImageDataset.create()
 dataset.add_samples_from_coco_caption(
     annotations_json="datasets/coco_subset_128_images/captions_train2017.json",
     images_path="datasets/coco_subset_128_images/images",