+++ lightly_studio/src/lightly_studio/core/dataset_query/field.py
@@ -9,8 +9,8 @@ from typing import Any, Generic, TypeVar, Union
 from sqlalchemy.orm import Mapped
 
 from lightly_studio.core.dataset_query.field_expression import (
+    ComparableFieldExpression,
     OrdinalFieldExpression,
-    StringFieldExpression,
 )
 
 T = TypeVar("T")
@@ -80,15 +80,15 @@ NumericalField = OrdinalField[Union[float, int]]
 DatetimeField = OrdinalField[datetime]
 
 
-class StringField(Field):
-    """Field for string values that supports equality operations.
+class ComparableField(Field, Generic[T]):
+    """Field for values that supports equality operations.
 
     Optional refactor when needed: Split into
-    - StringField(ABC) with the comparison operators.
-    - StringColumnField(StringField) for the __init__ and get_sqlmodel_field implementation.
+    - ComparableField(ABC) with the comparison operators.
+    - ComparableColumnField(StringField) for the __init__ and get_sqlmodel_field implementation.
     """
 
-    def __init__(self, column: Mapped[str]) -> None:
+    def __init__(self, column: Mapped[T]) -> None:
         """Initialize the string field with a database column.
 
         Args:
@@ -96,7 +96,7 @@ class StringField(Field):
         """
         self._column = column
 
-    def get_sqlmodel_field(self) -> Mapped[str]:
+    def get_sqlmodel_field(self) -> Mapped[T]:
         """Get the string database column or property.
 
         Returns:
@@ -104,10 +104,10 @@ class StringField(Field):
         """
         return self._column
 
-    def __eq__(self, other: str) -> StringFieldExpression:  # type: ignore[override]
+    def __eq__(self, other: T) -> ComparableFieldExpression[T]:  # type: ignore[override]
         """Create an equality expression."""
-        return StringFieldExpression(field=self, operator="==", value=other)
+        return ComparableFieldExpression(field=self, operator="==", value=other)
 
-    def __ne__(self, other: str) -> StringFieldExpression:  # type: ignore[override]
+    def __ne__(self, other: T) -> ComparableFieldExpression[T]:  # type: ignore[override]
         """Create a not-equal expression."""
-        return StringFieldExpression(field=self, operator="!=", value=other)
+        return ComparableFieldExpression(field=self, operator="!=", value=other)